added: separate ResponsiveImagesHelper function
